Flytter VedtakTilsynBarn for å kunne rename og bruke som en fellesklasse
