+++ src/main/kotlin/no/nav/tilleggsstonader/sak/vedtak/barnetilsyn/dto/AvslagTilsynBarnDto.kt
@@ -1,7 +1,7 @@
 package no.nav.tilleggsstonader.sak.vedtak.barnetilsyn.dto
 
 import no.nav.tilleggsstonader.sak.vedtak.TypeVedtak
-import no.nav.tilleggsstonader.sak.vedtak.barnetilsyn.ÅrsakAvslag
+import no.nav.tilleggsstonader.sak.vedtak.domain.ÅrsakAvslag
 
 data class AvslagTilsynBarnDto(
     val årsakerAvslag: List<ÅrsakAvslag>,
