+++ src/main/kotlin/no/nav/tilleggsstonader/sak/vedtak/barnetilsyn/dto/OpphørTilsynBarnDto.kt
@@ -1,7 +1,7 @@
 package no.nav.tilleggsstonader.sak.vedtak.barnetilsyn.dto
 
 import no.nav.tilleggsstonader.sak.vedtak.TypeVedtak
-import no.nav.tilleggsstonader.sak.vedtak.barnetilsyn.ÅrsakOpphør
+import no.nav.tilleggsstonader.sak.vedtak.domain.ÅrsakOpphør
 
 data class OpphørTilsynBarnDto(
     val årsakerOpphør: List<ÅrsakOpphør>,
